+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # v0.1.0
 
 - Intial release.
+
+# v0.2.1
+
+- Options extensions value change `[ext]` into `.[ext]`.
+- Fix ios android platform bug.